+++ src/net/sf/mpxj/primavera/PrimaveraReader.java
@@ -162,7 +162,6 @@ final class PrimaveraReader
          properties.setFinishDate(row.getDate("plan_end_date"));
          properties.setName(row.getString("proj_short_name"));
          properties.setStartDate(row.getDate("plan_start_date")); // data_date?
-         properties.setProjectTitle(row.getString("proj_short_name"));
          properties.setDefaultTaskType(TASK_TYPE_MAP.get(row.getString("def_duration_type")));
          properties.setStatusDate(row.getDate("last_recalc_date"));
          properties.setFiscalYearStartMonth(row.getInteger("fy_start_month_num"));
@@ -574,7 +573,22 @@ final class PrimaveraReader
     */
    public void processTasks(List<Row> wbs, List<Row> tasks, List<Row> udfVals)
    {
+      ProjectProperties projectProperties = m_project.getProjectProperties();
+      String projectName = projectProperties.getName();
       Set<Integer> uniqueIDs = new HashSet<Integer>();
+
+      //
+      // We set the project name when we read the project properties, but that's just
+      // the short name. The full project name lives on the first WBS item. Rather than
+      // querying twice, we'll just set it here where we have access to the WBS items.
+      // I haven't changed what's in the project name attribute as that's the value
+      // MPXJ users are used to receiving in that attribute, so we'll use the title
+      // attribute instead.
+      //
+      if (!wbs.isEmpty())
+      {
+         projectProperties.setProjectTitle(wbs.get(0).getString("wbs_name"));
+      }
 
       //
       // Read WBS entries and create tasks.
@@ -583,7 +597,7 @@ final class PrimaveraReader
       for (Row row : wbs)
       {
          Task task = m_project.addTask();
-         task.setProject(m_project.getProjectProperties().getName()); // P6 task always belongs to project
+         task.setProject(projectName); // P6 task always belongs to project
          processFields(m_wbsFields, row, task);
          uniqueIDs.add(task.getUniqueID());
          m_eventManager.fireTaskReadEvent(task);
@@ -632,7 +646,7 @@ final class PrimaveraReader
          {
             task = parentTask.addTask();
          }
-         task.setProject(m_project.getProjectProperties().getName()); // P6 task always belongs to project
+         task.setProject(projectName); // P6 task always belongs to project
 
          processFields(m_taskFields, row, task);
 
@@ -671,7 +685,7 @@ final class PrimaveraReader
          Date endDate = row.getDate("act_end_date") == null ? row.getDate("reend_date") : row.getDate("act_end_date");
          task.setFinish(endDate);
 
-         Duration work = Duration.add(task.getActualWork(), task.getRemainingWork(), m_project.getProjectProperties());
+         Duration work = Duration.add(task.getActualWork(), task.getRemainingWork(), projectProperties);
          task.setWork(work);
 
          // Add User Defined Fields